restructure the plugin
first off, things didn't work at all. The new sdk library was returning raw responses and we needed to hoist the stuff we actually care about from the other bits

second, we needed to set up the module exports so that it defaults to a thing that corsica can usehe but if you need to get after the business logic, as the tests do, then you can import that named export directly.

This will mean that we are not testing the bits tightly coupled to corsica. No mocking that beast or ensuring we've hooked into or parsed the messages correctly. Extract those things to custom functions if you feel so inclined. This way has the nice effect of isolating the stuff that changed and therefor the high risk bits of the release.

I'd add tests for the rest as we, or if we, ever need to modify it.